Create QuantumSecureDIDs.md
To address the risk of the advent of quantum computers for decentralized identity solutions, we propose to introduce a simple method using one-time signing keys and key rotation to protect our digital identity while using existing cryptographic ciphers for signing and hashing. Sam Smith's KERI is a potential candidate for implementing our proposed method. This approach shall allow us already today to prepare for the age of quantum attacks on our identity infrastructure.